Move duplicated code and remove dead code slugify
