+++ src/components/WhatsAppLink.jsx
@@ -1,16 +1,6 @@
 import { useState, useCallback } from 'react'
 import { useLang } from '../contexts/LanguageContext'
-
-function generateChallenge() {
-  const a = Math.floor(Math.random() * 9) + 1
-  const b = Math.floor(Math.random() * 9) + 1
-  const ops = [
-    { question: `${a} + ${b}`, answer: a + b },
-    { question: `${a + b} - ${b}`, answer: a },
-    { question: `${a} × ${b}`, answer: a * b },
-  ]
-  return ops[Math.floor(Math.random() * ops.length)]
-}
+import { generateChallenge } from '../utils/captcha'
 
 export function CaptchaModal({ href, onClose, t }) {
   const [challenge, setChallenge] = useState(generateChallenge)
